# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ghar Ka Sathi Production APIs",
    "version": "1.0.0",
    "description": "Comprehensive interactive API documentation for customer, expert, and admin application workflows. Built to support 80+ endpoints."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/auth/send-otp": {
      "post": {
        "tags": ["2. Authentication"],
        "summary": "Send OTP SMS",
        "description": "Sends verification OTP to mobile.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mobile": { "type": "string", "example": "9876543210" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OTP Sent" }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "tags": ["2. Authentication"],
        "summary": "Verify OTP & Generate JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mobile": { "type": "string", "example": "9876543210" },
                  "otp": { "type": "string", "example": "123456" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token generated" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["2. Authentication"],
        "summary": "Refresh Access Token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": { "type": "string", "example": "refresh_123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "New access token" }
        }
      }
    },
...
    "/address": {
      "post": {
        "tags": ["4. Address Management"],
        "summary": "Add Address to Book",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "house": { "type": "string", "example": "Flat 302" },
                  "building": { "type": "string", "example": "Green Gardens" },
                  "landmark": { "type": "string", "example": "Opp petrol pump" },
                  "latitude": { "type": "number", "example": 28.5355 },
                  "longitude": { "type": "number", "example": 77.3910 },
                  "tag": { "type": "string", "example": "Home" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Saved" }
        }
      },
      "get": {
        "tags": ["4. Address Management"],
        "summary": "Get All Address Book",
        "responses": {
          "200": { "description": "Address items array" }
        }
      }
    },
    "/address/{id}": {
      "put": {
        "tags": ["4. Address Management"],
        "summary": "Update Saved Address details",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Updated" }
        }
      },
      "delete": {
        "tags": ["4. Address Management"],
        "summary": "Delete Address item",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted" }
        }
      }
    },
    "/address/default": {
      "patch": {
        "tags": ["4. Address Management"],
        "summary": "Set Default Address indicator",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "addressId": { "type": "string", "example": "addr_123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Default updated" }
        }
      }
    },
...
  }
}